requirements.txt
Django==2.0.3
django-filter==2.0.0
django-crispy-forms==1.7.2